pyproject.toml
[project]
name = "tc-fiap-fase5"
version = "0.1.0"
description = "Projeto de Datathon - FIAP 2026"
authors = [
    {name = "Márcio Lima",email = "marciocad@gmail.com"},
    {name = "Luca Poit",email = "lucapoit@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.2",
    "mlflow==3.10.1",
    "pandera>=0.19.3",
    "pyarrow>=16.0.0",
    "pyyaml>=6.0.1",
    "joblib>=1.4.2",
    "dvc[gdrive] (>=3.67.1,<4.0.0)",
    "pytest>=8.2.0",
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "uvicorn>=0.30.0",
    "graphene==3.4.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "taskipy (>=1.14.1,<2.0.0)",
    "pyopenssl (>=23.0.0)",
    "cryptography (>=41.0.0)",
    "evidently>=0.7.21,<0.8",
    "rich>=13.9.0,<14",
    "xgboost>=2.1.0,<3",
    "prometheus-client>=0.22.1,<0.23",
]

[tool.poetry]
packages = [
    { include = "agent", from = "src" },
    { include = "common", from = "src" },
    { include = "features", from = "src" },
    { include = "models", from = "src" },
    { include = "monitoring", from = "src" },
    { include = "scenario_analysis", from = "src" },
    { include = "security", from = "src" },
    { include = "serving", from = "src" },
]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.1.0"
ruff = "^0.15.9"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
mlfeateng = 'python -m src.features.feature_engineering'
mltrain = 'python -m src.models.train'
mlscenarios = 'python -m src.scenario_analysis.inference_cases --suite-file configs/scenario_analysis/inference_cases.yaml'
mldrift = 'python -m src.monitoring.drift'
mldriftdemo = 'python -m src.monitoring.drift --current data/processed/test.parquet'
mlsyntheticdrift = 'python -m src.scenario_analysis.synthetic_drifts --all'
mlretrain = 'python -m src.models.retraining'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'
mlflow = 'mlflow server --backend-store-uri file:./mlruns --default-artifact-root ./mlruns --host 127.0.0.1 --port 5000'
serving = 'fastapi dev src/serving/app.py'
# Sobe containers com as imagens ja existentes (rapido). Apos mudar codigo em src/ ou Dockerfile, rode task observability_rebuild.
observability = 'docker compose up -d ollama ollama-pull serving mlflow prometheus grafana'
# Reconstrói a imagem da app (COPY src) e sobe a stack — use quando a API estiver desatualizada no container.
observability_rebuild = 'docker compose up -d --build ollama ollama-pull serving mlflow prometheus grafana'
observability_down = 'docker compose down'
# Lista modelos no Ollama (container tc-fiap-ollama ou, se falhar, host 127.0.0.1:11434).
ollama_list = 'python scripts/list_ollama_models.py'

pre_mlrunall = 'task mlfeateng'
mlrunall = 'python -m src.models.train && python -m src.models.train --config configs/training/experiments/random_forest_v1.yaml && python -m src.models.train --config configs/training/experiments/random_forest_v2.yaml && python -m src.models.train --config configs/training/experiments/random_forest_v3.yaml && python -m src.models.train --config configs/training/experiments/xgboost_v1.yaml && python -m src.models.train --config configs/training/experiments/gradient_boosting_v1.yaml && python -m src.scenario_analysis.inference_cases --suite-file configs/scenario_analysis/inference_cases.yaml'
